+++ src/ui/GalleryLayer.hpp
@@ -29,7 +29,7 @@ protected:
 	virtual bool init() override;
 	virtual void keyBackClicked() override;
 
-	CCSprite * getModeSprite(int tag, bool active);
+	CCSprite *getModeSprite(int tag, bool active);
 	void createModeButton(int tag, bool active = 0);
 	void setIDPopupClosed(SetIDPopup *popup, int value) override;
 
@@ -38,8 +38,6 @@ protected:
 	void fetchGallery();
 	void loadGallery();
 	matjson::Value m_fetchedData;
-
-	
 
 public:
 	unsigned int m_page = 0;
@@ -56,8 +54,13 @@ public:
 	void onFindPage(CCObject *sender);
 	void onSettings(CCObject *sender);
 	void onNavButton(CCObject *sender);
+	void onWebsite(CCObject *sender);
+	void onDiscord(CCObject *sender);
 
 	void refreshGallery();
+	// void showPackPopup();
+	// void createTexturePack();
+	void setupIconPack();
 
 	static GalleryLayer *create();
 	CCScene *scene();